+++ mastermind/src/metagame/AutoCodeCreater.py
